# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/drops/cron/transition": {
      "post": {
        "operationId": "DropsController_runCronTransition",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Transitions applied."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vercel Cron: transition drop statuses (requires CRON_SECRET)",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/discover": {
      "get": {
        "operationId": "DropsController_discover",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovery data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverDropsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get drops for the discover page (public-safe fields only)",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops": {
      "post": {
        "operationId": "DropsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDropDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drop"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/mine": {
      "get": {
        "operationId": "DropsController_getMyDrops",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Drop"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated user's drops",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/activity/mine": {
      "get": {
        "operationId": "DropsController_getMyActivity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DropActivityLog"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated user's activity logs across all drops",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/join/{joinCode}": {
      "get": {
        "operationId": "DropsController_findByJoinCode",
        "parameters": [
          {
            "name": "joinCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drop"
                }
              }
            }
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Look up a drop by join code",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}": {
      "get": {
        "operationId": "DropsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drop"
                }
              }
            }
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a drop by id (public access for public drops)",
        "tags": [
          "drops"
        ]
      },
      "patch": {
        "operationId": "DropsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDropDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drop"
                }
              }
            }
          },
          "400": {
            "description": "Drop is completed and cannot be edited."
          },
          "403": {
            "description": "Only the organiser can edit this drop."
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a drop (organiser only, active/ongoing status)",
        "tags": [
          "drops"
        ]
      },
      "delete": {
        "operationId": "DropsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Drop deleted successfully."
          },
          "403": {
            "description": "Only the organiser can delete this drop."
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a drop (organiser only)",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/activity": {
      "get": {
        "operationId": "DropsController_getActivityLogs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityLogsPageDto"
                }
              }
            }
          },
          "404": {
            "description": "Drop not found or no access."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated activity logs for a drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/crew/me": {
      "get": {
        "operationId": "DropsController_getMyCrewStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinDropResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not a crew member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current user's crew status for a drop",
        "tags": [
          "drops"
        ]
      },
      "delete": {
        "operationId": "DropsController_leaveDrop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully left the drop."
          },
          "403": {
            "description": "Organiser cannot leave their own drop."
          },
          "404": {
            "description": "Drop or crew membership not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave a drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/crew/me/presence": {
      "patch": {
        "operationId": "DropsController_updatePresence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresenceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Presence updated."
          },
          "400": {
            "description": "Not an active crew member."
          },
          "404": {
            "description": "Not a crew member of this drop."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark yourself in or out for a drop (active crew members only)",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/invite/{userId}": {
      "post": {
        "operationId": "DropsController_inviteToDrop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User invited successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a user to a drop (organiser only)",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/join": {
      "post": {
        "operationId": "DropsController_joinDrop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinDropResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Drop is completed."
          },
          "403": {
            "description": "Organiser cannot join their own drop, age requirement not met, or birthday missing for an age-restricted drop."
          },
          "404": {
            "description": "Drop or user not found."
          },
          "409": {
            "description": "Already joined this drop."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a drop",
        "tags": [
          "drops"
        ]
      }
    },
    "/drops/{id}/crew": {
      "get": {
        "operationId": "DropsController_getDropCrew",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrewMemberDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only the organiser can view the crew list."
          },
          "404": {
            "description": "Drop not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all crew members for a drop (organiser only)",
        "tags": [
          "drops"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DiscoverDropsResponseDto": {
        "type": "object",
        "properties": {
          "featured": {
            "nullable": true,
            "description": "Featured drop when available.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DropDiscoverSummaryDto"
              }
            ]
          },
          "recentChiefsDrops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DropDiscoverSummaryDto"
            }
          },
          "allPublic": {
            "$ref": "#/components/schemas/DiscoverDropsPaginatedDto"
          }
        },
        "required": [
          "recentChiefsDrops",
          "allPublic"
        ]
      },
      "CreateDropDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Beach Sunset Shoot"
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-05-10T18:00:00.000Z"
          },
          "location": {
            "type": "string",
            "example": "Sunset Beach, Manila"
          },
          "expectedHeadcount": {
            "type": "number",
            "example": 20
          },
          "overview": {
            "type": "string",
            "example": "This is a fun beach shoot"
          },
          "isLocked": {
            "type": "boolean",
            "example": false
          },
          "isPublic": {
            "type": "boolean",
            "example": true
          },
          "category": {
            "type": "string",
            "enum": [
              "hangout",
              "party"
            ],
            "example": "hangout"
          },
          "minimumAge": {
            "type": "number",
            "nullable": true,
            "description": "Minimum attendee age (≥). Only stored when category is party."
          },
          "idempotencyKey": {
            "type": "string",
            "example": "uuid-string"
          },
          "coverPhotoBase64": {
            "type": "string",
            "description": "Optional JPG/PNG cover as data URL; max 5MB decoded size."
          }
        },
        "required": [
          "name",
          "scheduledAt",
          "location"
        ]
      },
...
    }
  }
}